refactor: class-based plugins
The functional plugins were the source of too many user mistakes (by me!). Basically, some of the plugins had factory functions, some didn't, but they're all functions. Changing the plugins to be a known interface, and implementing them as classes eliminates the mistakes. It also reduces the memory footprint of creating mulitple clients or services because a lot of the shared functionality is on the prototype.